Fixed the scrolling issue - still a flicker
So I realized I was writing to the display after every character came in from the timer, not after a whole string was built.  So I fixed that, and implemented brightness adjustment via the potentiometer on A0.

I also did some clunky parsing to add the decimal, etc.

I'll send a video, but there's some flicker both while the timer is running, and when it's stopped, but I can't tell what's appearing on screen it's so quick.  But we're almost there!

Thanks again!